pyproject.toml
[project]
name = "flashcards-server"
version = "0.1.0"
description = "Flashcards API Server"
requires-python = ">=3.8"

[tool.poetry.dependencies]
python = "^3.8"
django = "5.1.5"
djangorestframework = "3.14.0"
django-cors-headers = "4.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 